Handle command output without blocking
Remove use of bufio.Scanner in favor of just writing to a bytes.Buffer
object. When a termination signal is received on the "killAll" channel
wait for the managed process to exit before reading buffer output.

The stdout test emits 128K + 1 bytes now in order to trigger a test
failure with the old implementation where scanner.Scan() would consume
the first 64K bytes, and the stdout pipe will be filled with the next
64K bytes leading a blocking condition by the next byte.  This is fixed
by the current implementation.

[#160826467]