REDVM-195: use azure go library
this moves away from blob xfer in favor
of the golang library.

the golang library includes it's own
retry logic for failed uploads and therefore
the timeout had to be increased in order
to detect failures.

azure unit test was removed because
it locked us into using blobxfer which
is no longer used